Prediction/requirements.txt
influxdb-client==1.40.0
flask==3.0.2
pandas==2.2.1
python-dotenv==1.0.1
scikit-learn==1.2.2
scipy==1.13.0
joblib==1.3.2